requirements.txt
selenium~=4.31.0
pytest~=8.3.5
python-dotenv~=1.1.0
allure-pytest
